fix(admin): move isScanWithoutIndex out of route.ts for Next route-export rules
Next.js App Router route files may only export HTTP handlers + route config;
exporting the `isScanWithoutIndex` helper from route.ts failed `next build`
route-export validation (tsc does not catch this — only the production build
does). Relocated the pure helper to a sibling `scan-detection.ts` and updated
the route + test imports. 11 tests pass; next build compiles clean.
