+++ app/api/admin/schema-health/route.ts
@@ -26,34 +26,7 @@ import { NextRequest, NextResponse } from "next/server";
 import { getCloudflareContext } from "@opennextjs/cloudflare";
 import { requireAdmin } from "@/lib/admin/auth";
 import { LEADERBOARD_AGGREGATE_SQL } from "@/lib/competition/leaderboard-query";
-
-// ---------------------------------------------------------------------------
-// SCAN detection
-// ---------------------------------------------------------------------------
-
-/**
- * Determine whether an EXPLAIN QUERY PLAN detail line represents a full table
- * scan without any index.
- *
- * SQLite EXPLAIN QUERY PLAN output shapes:
- *   "SCAN tablename"                                    → BAD  (full scan, no index)
- *   "SCAN tablename USING COVERING INDEX idx_name"      → OK   (index-only scan)
- *   "SCAN tablename USING INDEX idx_name"               → OK   (index scan)
- *   "SEARCH tablename USING INDEX idx_name (...)"       → OK   (index seek)
- *   "SEARCH tablename USING COVERING INDEX idx_name"    → OK   (covering index seek)
- *   "USE TEMP B-TREE FOR ORDER BY"                      → informational, not a scan
- *
- * We flag only the first case: SCAN without any USING INDEX clause.
- */
-export function isScanWithoutIndex(detail: string): boolean {
-  // Must start with "SCAN " (SQLite uses uppercase)
-  if (!detail.includes("SCAN ")) return false;
-  // If it uses any index, it is acceptable
-  if (detail.includes("USING INDEX") || detail.includes("USING COVERING INDEX")) {
-    return false;
-  }
-  return true;
-}
+import { isScanWithoutIndex } from "./scan-detection";
 
 // ---------------------------------------------------------------------------
 // Query definitions
